CRITICAL FIX: Resolve FlashInfer RoPE mixed precision data type requirements
- Fix cos_sin_cache data type from fp16 to float32 across all models
- FlashInfer requires cos_sin_cache to be float32 while query/key can be fp16
- Update DeepSeek model RoPE cache initialization to use torch.float32
- Update LLaMA model _set_cos_sin_cache to use torch.float32
- Update Qwen model _set_cos_sin_cache to use torch.float32
- Fix mla_attn.py cos_sin_cache conversion for FlashInfer compatibility

Resolves "cos_sin_cache should be float32" error in FlashInfer RoPE API calls.

+++ model_hub/llama.py
@@ -138,8 +138,8 @@ class LlamaModel(LLM):
         - 缓存大小基于max_length而非max_position_embeddings
         - 应用了attention_scaling以支持长序列
         """
-        t = torch.arange(self.max_length, device=self.inv_freq.device, dtype=self.dtype)  # 使用模型数据类型
-        freqs = torch.outer(t, self.inv_freq.to(self.dtype))  # 确保inv_freq也是正确类型
+        t = torch.arange(self.max_length, device=self.inv_freq.device, dtype=torch.float32)  # FlashInfer要求float32
+        freqs = torch.outer(t, self.inv_freq.to(torch.float32))  # FlashInfer要求float32
         return freqs.cos()*self.attention_scaling, freqs.sin()*self.attention_scaling
